+++ src/dfx/src/util/clap/validators.rs
@@ -38,7 +38,7 @@ pub fn memo_validator(memo: &str) -> Result<(), String> {
 }
 
 pub fn cycle_amount_validator(cycles: &str) -> Result<(), String> {
-    if cycles.parse::<u64>().is_ok() {
+    if cycles.parse::<u128>().is_ok() {
         return Ok(());
     }
     Err("Must be a non negative amount.".to_string())
